backend/requirements.txt
# 移动账本后端依赖
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.25
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pytz==2025.2
httpx==0.27.0
bcrypt==4.2.0
